enable -sub output
